+++ src/permissions-mask.ts
@@ -1,0 +1,2 @@
+export type PermissionsMask = Map<PermissionDescriptor, PermissionMask>;
+export type PermissionMask = Partial<Record<PermissionState, PermissionState>>;